cybersecurity_demo/requirements.txt
# Core Framework
Flask==3.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# Authentication
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Rate Limiting
Flask-Limiter==3.5.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Vector Store & Embeddings
numpy<2.0
chromadb==0.4.22
sentence-transformers==2.3.1

# AI/LLM
requests==2.31.0

# Utilities
python-dotenv==1.0.0
werkzeug==3.0.1

# Production Server
gunicorn==21.2.0
eventlet==0.33.3
